+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## master (unreleased)
 ### Security
 ### Added
+- `Mail::Message#extract_html` extracts the html part as a counterpart for `Mail::Message#extract_text`. This is useful when an email has no text part.
+- `Parsers::HtmlMails#parse` removes quotes indicated by `<div name="quote"></div>`.
 ### Deprecated
 ### Removed
 ### Fixed